Removed all the TODOs, FIXMEs, and finally renamed newGCSXbee to GCSXbee
