+++ GCSXbee.py
@@ -1,314 +1,379 @@
-import csv
 import socket
-import os
 import time
-from threading import Thread
+import threading
+import telemetryPacket
 from datetime import datetime, timezone
-
-# This is because there is a funny issue where some package requires a yaml config.
-# I have no clue what causes this or how to fix it, sometimes it is needed and othertimes it doesn't.
-# When I was developing the code, my computer was the only one that seemed to be in the othertimes camp always.
-# Only God knows why this is.
-# os.environ["KUBECONFIG"] = os.path.abspath("kube_config.yaml")
-
 from digi.xbee.devices import XBeeDevice, RemoteXBeeDevice, XBee64BitAddress
+import enum
 
-address = ('localhost', 6000)
+HOST = 'localhost'
+PORT_TH = 6000
+PORT_GUI = 6001
+ADDRESS_TH = (HOST, PORT_TH)
+ADDRESS_GUI = (HOST, PORT_GUI)
+DEFAULT_PRESSURE = 101325
+
+# Colors
+CRED = '\033[91m'
+CGREEN = '\033[92m'
+CYELLOW = '\033[93m'
+CBLUE = '\033[94m'
+CCYAN = '\033[96m'
+CRESET = '\033[0m'
+
+class Status(enum.Enum):
+    DISABLED = enum.auto()
+    WAITING_ENABLE = enum.auto()
+    ENABLED = enum.auto()
+    WAITING_ACTIVE = enum.auto()
+    ACTIVE = enum.auto()
+    WAITING_DISABLED = enum.auto()
 
 class TelemetryHandler:
-    def __init__(self, team_id, port="COM3", baudrate=9600, write_path=None, mac_addr="0013A20041E060D1"): # default port val for Fernando's laptop
-        """
-        Initialize the telemetry handler.
 
-        Args:
-            team_id (str): Four-digit team identification number.
-            port (str): Serial port for communication.
-            baudrate (int): Baud rate for XBee communication.
-            mac_addr (str): Mac Address of the XBee on the Can, part of the unicast prevention.
-        """
+    def __init__(self, team_id, xbee_port, xbee_baudrate, xbee_target_mac_addr, press_csv_path=None, log_csv_path=None):
 
-        ##################### File path for the simulated data to be used #####################
-        self.SIM_CSV_PATH = "SIM_Pressure.csv"  # Path to the simulated data CSV file
-        #######################################################################################
-
-        # Some of these definitions are redundant, but they are here for clarity
+        # Load some of the init variables
         self.team_id = team_id
-        self.is_receiving = False
-        self.csv_file = None
-        self.csv_writer = None
-        self.packet_count = 0
-        self.sim_enable = False
-        self.sim_activate = False
-        self.simulation_thread = None
-        self.receiver = None
-        self.write_filepath = write_path
-        if self.write_filepath == None:
-            raise Exception(f"GCSXbee (File: GCSXbee.py Function: __init__) [INITIALIZATION] : No file write_path given")
+        self.press_csv_path = press_csv_path
+        self.log_csv_path = log_csv_path
 
-        # Define telemetry fields as per competition requirements
-        self.telemetry_fields = ["TEAM_ID","MISSION_TIME","PACKET_COUNT","MODE","STATE","ALTITUDE",
-              "TEMPERATURE", "PRESSURE", "VOLTAGE","CURRENT",
-              "GYRO_R", "GYRO_P", "GYRO_Y", "ACCEL_R", "ACCEL_P", "ACCEL_Y",
-              "GPS_TIME", "GPS_ALTITUDE", "GPS_LATITUDE", "GPS_LONGITUDE", "GPS_SATS", "CMD_ECHO"]
+        # Operation Variables
+        self.is_running = True
+        self.sim_status = Status.DISABLED
+        self.sim_disable_pending = False
+        self.latest_pkt = False
+        self.valid_xbee_connection = False
+        self.valid_pressure_file = False
+        self.valid_log_file = False
 
         # Initialize XBee connection
-        self.mac_address = mac_addr # This is the MAC address of the FSW radio (the one on the Sat)
-        self.xbee_device = XBeeDevice(port, baudrate)
-        self.receiver = RemoteXBeeDevice(x64bit_addr=XBee64BitAddress.from_hex_string(self.mac_address), local_xbee=self.xbee_device)
-        # FIXME : This MAC address will need to be updated to the actual FSW radio's MAC address
+        self.xbee_port = xbee_port
+        self.xbee_baudrate = xbee_baudrate
+        self.xbee_target_mac_addr = xbee_target_mac_addr
+        self.xbee_device = XBeeDevice(self.xbee_port, self.xbee_baudrate)
+        self.xbee_receiver = RemoteXBeeDevice(x64bit_addr=XBee64BitAddress.from_hex_string(self.xbee_target_mac_addr), local_xbee=self.xbee_device)
 
     def start_telemetry(self):
-        """Start receiving telemetry data."""
+
+        print("[DEBUG] Handler Started")
+
+        # Open the xbee device
         try:
             self.xbee_device.open()
+            self.valid_xbee_connection = True
+            print(f"[DEBUG] XBee Device Open Successful: {self.xbee_port}")
         except Exception as e:
-            raise Exception(f"GCSXbee (File: GCSXbee.py Function: start_telemetry) [START TELEMETRY] Failed to open XBee device: {e}")
-            return # Prevents the .csv from being overwritten. Should have done this sooner.
+            print(f"[DEBUG] Failed to open XBee device: {e}")
+            self.valid_xbee_connection = False
+
+        # Open the pressure file
+        try:
+            self.press_csv_file = open(self.press_csv_path,'rt')
+            self.valid_pressure_file = True
+            print(f"[DEBUG] Pressure CSV Open Successful: {self.press_csv_path}")
+        except Exception as e:
+            print(f"[DEBUG] Failed to Open Pressure CSV File, Defaulting to {DEFAULT_PRESSURE} Pascals: {e}")
+            self.press_csv_file = None
+            self.valid_pressure_file = False
+
+        # Open the log file
+        try:
+            self.log_csv_file = open(self.log_csv_path,"w", encoding="utf-8")
+            self.valid_log_file = True
+            print(f"[DEBUG] Log CSV Open Successful: {self.log_csv_path}")
+            self.log_csv_file.write(f"TEAM_ID,MISSION_TIME,PACKET_COUNT,MODE,STATE,ALTITUDE,TEMPERATURE,PRESSURE,VOLTAGE,CURRENT,GYRO_R,GYRO_P,GYRO_Y,ACCEL_R,ACCEL_P,ACCEL_Y,GPS_TIME,GPS_ALTITUDE,GPS_LATITUDE,GPS_LONGITUDE,GPS_SATS,CMD_ECHO,,\n")
+            self.log_csv_file.flush()
+        except Exception as e:
+            print(f"[DEBUG] Failed to Open Log CSV File: {e}")
+            self.log_csv_file = None
+            self.valid_log_file = False 
+
+        # Create the three threads, one for monitoring the incoming commands, one for running the main execution loop that sends data to the GUI, one for running the simulation loop when in simulation mode
+        # Seperate threads are necessary for each operation as they needs to be "Asynchronous" because they are time sensitive
+        self.forwarding_thread = threading.Thread(target=self._forwarding_loop, daemon=True)
+        self.command_thread = threading.Thread(target=self._command_loop, daemon=True)
+        self.simulation_thread = threading.Thread(target=self._simulation_loop, daemon=True)
         
-        # FIXME:
-        # Create CSV file with specified naming format
-        #self.csv_file = open(self.write_filepath, 'w', newline='')
-        #self.csv_writer = csv.writer(self.csv_file)
+        self.forwarding_thread.start()
+        self.command_thread.start()
+        self.simulation_thread.start()
 
-        # Write header row
-        #self.csv_writer.writerow(self.telemetry_fields)
-
-        # Start receiving data
-        self.is_receiving = True
-        self.receive_thread = Thread(target=self._receive_telemetry, daemon=True)
-        self.receive_thread.start()
+        return
 
     def stop_telemetry(self):
-        """Stop receiving telemetry data and close files."""
-        self.is_receiving = False
-        if self.receive_thread:
-            self.receive_thread.join()
-
-        # FIXME:
-        #if self.csv_file:
-        #    self.csv_file.close()
-
-        print(f"GCSXbee (File: GCSXbee.py Function: stop_telemetry) [STOP TELEMETRY] : Telemetry stopped. {self.packet_count} packets received.")
-
-        # if self.xbee_device and self.xbee_device.is_open():
-        #     self.send_command(f"CMD,{self.team_id},CX,OFF")
-        #     self.xbee_device.close()
-
-    def send_command(self, command):
-        """
-        Send a command to the CanSat.
-
-        Args:
-            command (str): Command string following competition format.
-        """
-
-        # Because the FSW uses a buffer to help read the commands sent to it, if the buffer is not filled right away it will wait until it is filled.
-        # This will cause some commands to need to be sent twice. Padding the strings here with null terms allows the commands to be sent once.
-        if command == "CX ON":
-            print("[DEBUG] CX-ON COMMAND ENTERED") #FIXME: REMOVE
-            CXON = f"CMD,{self.team_id},CX,ON\0\0\0\0\0\0\0\0"
-            try:
-                print("[DEBUG] CX-ON COMMAND INSIDE") #FIXME: REMOVE
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=CXON)
-                    print("[DEBUG] CX-ON COMMAND SUCCESS") #FIXME: REMOVE
-                else:
-                    print("[DEBUG] CX-ON COMMAND FAILED") #FIXME: REMOVE
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND CXON]: Error sending command - {e}")
-
-        elif command == "CX OFF":
-            print("[DEBUG] CX-OFF COMMAND ENTERED") #FIXME: REMOVE
-            CXOFF = f"CMD,{self.team_id},CX,OFF\0\0\0\0\0\0\0"
-            try:
-                print("[DEBUG] CX-OFF COMMAND INSIDE") #FIXME: REMOVE
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver ,data=CXOFF)
-                    print("[DEBUG] CX-OFF COMMAND SUCCESS") #FIXME: REMOVE
-                else:
-                    print("[DEBUG] CX-OFF COMMAND FAILED") #FIXME: REMOVE
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND CXOFF]: Error sending command - {e}")
         
-        elif command == "SIM ENABLE":
-            ENABLE = f"CMD,{self.team_id},SIM,ENABLE\0\0\0"
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=ENABLE)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND SIM ENABLE]: Error sending command - {e}")
+        self.is_running = False
 
-        elif command == "SIM ACTIVATE":
-            ACTIVATE = f"CMD,{self.team_id},SIM,ACTIVATE\0"
-            try:
-                if self.xbee_device.is_open() and self.sim_enable:
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=ACTIVATE)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND SIM ACTIVATE]: Error sending command - {e}")
-        
-        elif command == "SIM DISABLE":
-            DISABLE = f"CMD,{self.team_id},SIM,DISABLE\0\0"
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=DISABLE)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND SIM DISABLE]: Error sending command - {e}")
-
-        elif command == "CAL":
-            CAL = f"CMD,{self.team_id},CAL\0\0\0\0\0\0\0\0\0\0"
-            print(f"Sending command: {CAL}")
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=CAL)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND CAL]: Error sending command - {e}")
-
-        elif command == "ST GPS":
-            ST_GPS = f"CMD,{self.team_id},ST,GPS\0\0\0\0\0\0\0"
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=ST_GPS)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND ST GPS]: Error sending command - {e}")
-
-        elif command[0:2] == "ST":
-            current_time = "00:00:00"
-            try: # This is just a bunch of handling I did in case the user gives an incomplete cmd, probably not necessary but is nice to have.
-                current_time = command[3:]
-                if current_time == "":
-                    current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-                elif len(current_time) != 8:
-                    current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-                elif current_time.count(":") != 2:
-                    current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-                elif current_time[2] != ":" or current_time[5] != ":":
-                    current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-                elif int(current_time[0:2]) > 23 or int(current_time[3:5]) > 59 or int(current_time[6:8]) > 59:
-                    current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-            
-            except:
-                current_time = datetime.now(timezone.utc).strftime('%H:%M:%S') # Get the current time in UTC
-            print(current_time)
-            ST = f"CMD,{self.team_id},ST,{current_time}\0\0"
-            print(ST)
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=ST)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND ST]: Error sending command - {e}")
-
-        elif command == "MEC WIRE ON":
-            MEC_WIRE = f"CMD,{self.team_id},MEC,WIRE,ON\0\0"
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=MEC_WIRE)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND MEC WIRE ON]: Error sending command - {e}")
-
-        elif command == "MEC WIRE OFF":
-            MEC_WIRE = f"CMD,{self.team_id},MEC,WIRE,OFF\0"
-            try:
-                if self.xbee_device.is_open():
-                    self.xbee_device.send_data_async(remote_xbee=self.receiver, data=MEC_WIRE)
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: send_command) [COMMAND MEC WIRE OFF]: Error sending command - {e}")
-
-        # FIXME : Add any other MEC commands here -------------------------------------------------------------------------------
-
-        else:
-            print(f"ERROR (File: GCSXbee.py Function: send_command) [SEND_COMMAND]: Unknown command - {command}")
-
-    def _receive_telemetry(self):
-        """Internal method to receive and process telemetry data."""
-        while self.is_receiving:
-            try:
-                xbee_message = self.xbee_device.read_data(20) # The 20 is a timeout parameter
-                if xbee_message:
-                    # Read and decode the message
-                    line = xbee_message.data.decode('utf-8').strip()
-                    xbee_message = self.xbee_device.read_data(20)  # Read the next message
-                    if xbee_message is None:
-                        continue
-                    line = line + xbee_message.data.decode('utf-8').strip()  # Append the next message data
-                    #print(f"[RECEIVE TELEMETRY] Received - [{line}]") #FIXME: Remove this later
-                    data = line.split(',')
-
-                    # Because we could not get the GPS to work, we have to fake all of the GPS data.
-                    # This includes GPS time, but it is easier to do that here in the GCS. Whoopsies.
-                    #current_time = datetime.now(timezone.utc).strftime('%H:%M:%S')
-                    #data[19] = current_time
-
-                    # Validate team ID and basic data format
-                    if (len(data) >= len(self.telemetry_fields)) and (data[0] == self.team_id):
-                        
-                        # FIXME: DO some formatting later
-                        # Write to CSV file
-                        #self.csv_writer.writerow(data)
-                        #self.csv_file.flush()  # Ensure data is written to disk
-
-                        # Write to the socket instead 
-                        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-                        sock.connect(address)
-                        sock.send(line.encode('utf-8'))
-                        print("Sent:",line)
-                        sock.close()
-
-                        # Update packet count
-                        self.packet_count += 1
-
-                    # # FIXME : This may need to be updated to handle the format that FSW sends us (the array index that is) -------------------------
-                    # if data[24] == "SIMENABLE":
-                    #     self.sim_enable = True
-
-                    # elif data[24] == "SIMACT":
-                    #     self.sim_activate = True 
-                    #     self.start_sim()                     
-
-                    # elif data[24] == "SIMDIS":
-                    #     self.sim_activate = False
-                    #     self.sim_enable = False
-                    #     if self.simulation_thread:
-                    #         self.stop_sim()
-
-            except Exception as e:
-                print(f"ERROR (File: GCSXbee.py Function: _receive_telemetry) [RECEIVE TELEMETRY] : {e}")
-
-    def start_sim(self):
-        if (self.sim_enable and self.sim_activate):
-            print(self.SIM_CSV_PATH)
-            self.simulation_thread = Thread(target=self._send_command_pressure)
-            self.simulation_thread.start()
-
-    def stop_sim(self):
-        """
-        Stop sending simulated pressure data.
-        """
-        # while self.simulation_thread:
-        print("Waiting for simulation thread to finish...")
+        self.forwarding_thread.join()
+        self.command_thread.join()
         self.simulation_thread.join()
 
-        self.sim_enable = False
-        self.sim_activate = False
-        print("Simulation stopped.")
+        return
 
-    def _send_command_pressure(self):
-        """
-        Send simulated pressure data (simulation mode only).
+    def _command_loop(self):
 
-        Args:
-            csv_path (string): Path to the CSV file containing pressure data.
-        """
-        
-        with open(self.SIM_CSV_PATH, 'r') as csv_file:
-            csv_reader = csv.reader(csv_file)
-            for row in csv_reader:
-                if not self.sim_enable or not self.sim_activate: # Check if simulation is enabled and activated
-                    print("Simulation disabled or not activated. Stopping simulation thread.")
-                    break
-                DATA = f"CMD,{self.team_id},SIMP,{row[0]}"
-                self.xbee_device.send_data_async(remote_xbee=self.receiver, data=DATA)
-                time.sleep(1)
+        sock_in = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        sock_in.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        sock_in.bind(ADDRESS_TH)
+        sock_in.listen(0)
+        sock_in.settimeout(2)
 
-        DATA = f"CMD,{self.team_id},SIM,DISABLE";
-        self.xbee_device.send_data_async(remote_xbee=self.receiver, data=DATA)
-        print("Simulation thread finished.")
+        while (self.is_running):
+            try:
+                conn, addr = sock_in.accept()
+                full_data = ""
+                while True:
+                    data = conn.recv(1024)
+                    if not data: break
+                    full_data += data.decode('utf-8')
+                conn.close()
+                print("[DEBUG] TLH Recv:", full_data)
+                self._handle_command(full_data)
+            except socket.timeout:
+                #print("[DEBUG] Socket Timout")
+                1==1
+
+    def _forwarding_loop(self):
+
+        #time.sleep(1)
+        i = 1
+
+        while (self.is_running == True):
+            # Write to the socket instead 
+            time.sleep(1)
+            #print("[DEBUG] LOOP OUTGOING")
+
+            if self.valid_xbee_connection:
+                try:
+                    if self.xbee_device.is_open():
+                        xbee_message = self.xbee_device.read_data(20) # The 20 is a timeout parameter
+                        if xbee_message:
+                            line = xbee_message.data.decode('utf-8').strip()
+                            self.latest_pkt = telemetryPacket.TelemetryPacket(line)
+                            print(f"[DEBUG PACKET:]{line}")
+                            if (self.valid_log_file):
+                                print("[DEBUG] LOGGED THE PACKET")
+                                self.log_csv_file.write(f"{self.latest_pkt.get_str()}\n")
+                                self.log_csv_file.flush()
+                            print("")
+
+                            # Update the state of the simulator if it's in a pending state for any modes
+                            if self.latest_pkt.CMD_ECHO == "SIMDIS" and self.sim_status == Status.DISABLED and self.sim_disable_pending == True:
+                                self.sim_disable_pending = False
+                                print(f"{CYELLOW}[INFO] FSW SIMULATION DISABLE CONFIRMED{CRESET}")
+                            elif self.latest_pkt.CMD_ECHO == "SIMENABLE" and self.sim_status == Status.WAITING_ENABLE:
+                                self.sim_status = Status.ENABLED
+                                print(f"{CYELLOW}[INFO] SIMULATION NOW ENABLED{CRESET}")
+                            elif self.latest_pkt.CMD_ECHO == "SIMACT" and self.sim_status == Status.WAITING_ACTIVE:
+                                if self.latest_pkt.MODE == "S":
+                                    self.sim_status = Status.ACTIVE
+                                    print(f"{CGREEN}[INFO] SIMULATION NOW ACTIVE{CRESET}")
+                                else:
+                                    print("[ERROR] FSW RECEIVED ACTIVE REQUEST BUT MODE DID NOT CHANGE")
+                        else:
+                            self.latest_pkt = None
+                            print("[DEBUG] Xbee Device Read Timout. (Latest Packet is now \"None\")")
+                    else:
+                        print("[DEBUG] Packet Polling Failed, XBEE Closed")
+                except Exception as e:
+                    print(f"[ERROR] Xbee Device Read Error Occurred: {e}")
+                    self.latest_pkt = None
+                    #self.valid_xbee_connection = False
+
+                if (self.latest_pkt and self.latest_pkt.valid_packet):
+                    line = self.latest_pkt.get_str()
+                    sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+
+                    if (not sock.connect_ex(ADDRESS_GUI)):    
+                        sock.send(line.encode('utf-8'))
+                        #print("[DEBUG] TLH Sent:", line)
+                        i += 1
+                    else:
+                        print("[DEBUG] TLH Failed Send")
+                        1==1
+
+                    sock.close()
+
+    def _simulation_loop(self):
+
+        while (self.is_running == True):
+            # Write to the socket instead 
+            time.sleep(1)
+
+            while (self.sim_status == Status.ACTIVE):
+                if (self.valid_pressure_file):
+                    temp_pressure_str = self.press_csv_file.readline()
+                    temp_clean_str = temp_pressure_str.strip("\n")
+                    if temp_pressure_str:
+                        self._send_packet(f"CMD,1075,SIMP,{temp_clean_str}")
+                        print(f"{CCYAN}[DEBUG] Pressure Packet: {temp_clean_str}{CRESET}")
+                        time.sleep(1)
+                    else:
+                        print(f"{CBLUE}[DEBUG] End of Pressure File{CRESET}")
+                        time.sleep(1)
+                else:
+                    #print(f"[DEBUG] NO PRESSURE FILE OPEN")
+                    temp_clean_str = str(DEFAULT_PRESSURE)
+                    self._send_packet(f"CMD,1075,SIMP,{temp_clean_str}")
+                    print(f"{CCYAN}[DEBUG] Pressure Packet: {temp_clean_str}{CRESET}")
+                    time.sleep(1)
+
+        return
+
+    def _handle_command(self, cmd):
+
+        str = "[DEBUG] DUMMY COMMAND PACKET"
+        fields = cmd.split(",")
+        cmd_field = fields[0]
+        match (cmd_field):
+            case "CX_ON":
+                str = "CMD,1075,CX,ON"
+                self._send_packet(str)
+            case "CX_OFF":
+                str = "CMD,1075,CX,OFF"
+                self._send_packet(str)
+            case "ST_UTC":
+                utc_now = datetime.now(timezone.utc)
+                utc_time = utc_now.strftime("%H:%M:%S")
+                packet_str = f"CMD,1075,ST,{utc_time}"
+                self._send_packet(packet_str)
+            case "ST_GPS":
+                str = "CMD,1075,ST,GPS"
+                self._send_packet(str)
+            case "SIM_EN":
+                str = "CMD,1075,SIM,ENABLE"
+                self._send_packet(str)
+                match(self.sim_status):
+                    case Status.DISABLED:
+                        self.sim_status = Status.WAITING_ENABLE
+                        print("[DEBUG] SIMULATION ENABLE NOW PENDING")
+                    case Status.WAITING_ENABLE:
+                        print("[DEBUG] SIMULATION ENABLE ALREADY PENDING (RESENDING)")
+                    case Status.ENABLED:
+                        print("[DEBUG] SIMULATION ALREADY ACTIVE")
+                    case Status.WAITING_ACTIVE:
+                        print("[DEBUG] SIMULATION ALREADY ACTIVE")
+                    case Status.ACTIVE:
+                        print("[DEBUG] SIMULATION ALREADY ENABLED AND ACTIVE")
+                    case Status.WAITING_DISABLED:
+                        print("[DEBUG] CAN'T ENABLE, SIMULATION DISABLE PENDING")
+                    case _:
+                        print("[ERROR] SIM STATUS IN INVALID STATE")
+            case "SIM_ACT":
+                str = "CMD,1075,SIM,ACTIVATE"
+                self._send_packet(str)
+                match(self.sim_status):
+                    case Status.DISABLED:
+                        print("[DEBUG] SIMULATION IS OFF")
+                    case Status.WAITING_ENABLE:
+                        print("[DEBUG] SIMULATION ENABLE STILL PENDING")
+                    case Status.ENABLED:
+                        self.sim_status = Status.WAITING_ACTIVE
+                        print("[DEBUG] SIMULATION ACTIVE NOW PENDING")
+                    case Status.WAITING_ACTIVE:
+                        print("[DEBUG] SIMULATION ACTIVE ALREADY PENDING (RESENDING)")
+                    case Status.ACTIVE:
+                        print("[DEBUG] SIMULATION ALREADY ACTIVE")
+                    case Status.WAITING_DISABLED:
+                        print("[DEBUG] CAN'T ACTIVATE, SIMULATION DISABLE PENDING")
+                    case _:
+                        print("[ERROR] SIM STATUS IN INVALID STATE")
+            case "SIM_DIS":
+                if self.press_csv_file and self.valid_pressure_file:
+                    self.press_csv_file.seek(0) #Resets the the pressure file's position anytime simulation is disabled to allow reruns
+                str = "CMD,1075,SIM,DISABLE"
+                self._send_packet(str)
+                self.sim_disable_pending = True
+                match(self.sim_status):
+                    case Status.DISABLED:
+                        print("[DEBUG] SIMULATION IS ALREADY OFF")
+                    case Status.WAITING_ENABLE:
+                        self.sim_status = Status.DISABLED
+                        print("[DEBUG] SIMULATION DISABLE NOW PENDING, LOCAL SIM DISABLED")
+                    case Status.ENABLED:
+                        self.sim_status = Status.DISABLED
+                        print("[DEBUG] SIMULATION DISABLE NOW PENDING, LOCAL SIM DISABLED")
+                    case Status.WAITING_ACTIVE:
+                        self.sim_status = Status.DISABLED
+                        print("[DEBUG] SIMULATION DISABLE NOW PENDING, LOCAL SIM DISABLED")
+                    case Status.ACTIVE:
+                        self.sim_status = Status.DISABLED
+                        print("[DEBUG] SIMULATION DISABLE NOW PENDING, LOCAL SIM DISABLED")
+                    case _:
+                        print(f"[ERROR] SIM STATUS IN INVALID STATE {self.sim_status}, SET TO DISABLED")
+                        self.sim_status = Status.DISABLED
+            case "CAL":
+                str = "CMD,1075,CAL"
+                self._send_packet(str)
+            case "GNC_ACT":
+                str = "CMD,1075,GNC_ACT,"
+                self._send_packet(str)
+            case "MEC":
+                if len(fields) >= 3:
+                    str = f"CMD,1075,MEC,{fields[1]},{fields[2]}"
+                    #print(f"[DEBUG] MECH COMMAND: {fields}")
+                    self._send_packet(str)
+                else:
+                    print("[ERROR] INVALID MECH FORMAT RECEIVED")
+            # Fake commands
+            case "*REOPEN_RADIO":
+                if (self.xbee_device):
+                    self.xbee_device.close()
+                # Re-open the xbee device
+                try:
+                    self.xbee_device.open()
+                    self.valid_xbee_connection = True
+                    print(f"[DEBUG] XBee Device Re-open Successful: {self.xbee_port}")
+                except Exception as e:
+                    print(f"[DEBUG] Failed to Re-open XBee device: {e}")
+                    self.valid_xbee_connection = False
+            case "*REOPEN_PRESSURE":
+                if (self.press_csv_file):
+                    self.press_csv_file.close()
+                self.valid_pressure_file = False
+                # Re-open the pressure file
+                try:
+                    self.press_csv_file = open(self.press_csv_path,'rt')
+                    self.valid_pressure_file = True
+                    print(f"[DEBUG] Pressure CSV Re-open Successful: {self.press_csv_path}")
+                except Exception as e:
+                    print(f"[DEBUG] Failed to Re-open Pressure CSV File: {e}")
+                    self.valid_pressure_file = False
+            case "*REOPEN_LOG":
+                if (self.log_csv_file):
+                    self.log_csv_file.close()
+                self.valid_log_file = False
+                # Re-open the log file
+                try:
+                    self.log_csv_file = open(self.log_csv_path,"w", encoding="utf-8")
+                    self.valid_log_file = True
+                    print(f"[DEBUG] Log CSV Re-open Successful: {self.log_csv_path}")
+                    self.log_csv_file.write(f"TEAM_ID,MISSION_TIME,PACKET_COUNT,MODE,STATE,ALTITUDE,TEMPERATURE,PRESSURE,VOLTAGE,CURRENT,GYRO_R,GYRO_P,GYRO_Y,ACCEL_R,ACCEL_P,ACCEL_Y,GPS_TIME,GPS_ALTITUDE,GPS_LATITUDE,GPS_LONGITUDE,GPS_SATS,CMD_ECHO,,\n")
+                    self.log_csv_file.flush()
+                except Exception as e:
+                    print(f"[DEBUG] Failed to Re-open Log CSV File: {e}")
+                    self.valid_log_file = False 
+            case "*DEACTIVATE_PRESSURE":
+                if (self.press_csv_file):
+                    self.press_csv_file.close()
+                self.valid_pressure_file = False
+                print(f"[DEBUG] Closed Pressure CSV File, Defaulting to {DEFAULT_PRESSURE} Pascals")
+            case _:
+                print("[DEBUG] INVALID COMMAND RECEIVED")
+
+        return
+
+    def _send_packet(self, str):
+        #print("DEBUG BLAHHHHHHHHH")
+        if self.valid_xbee_connection:
+            #print("DEBUG BLEHHHHHHHHH")
+            try:
+                if self.xbee_device.is_open():
+                    #print("BLOHHHHHHH")
+                    self.xbee_device.send_data_async(remote_xbee=self.xbee_receiver, data=str)
+                    print(f"[DEBUG] CMD: \"{str}\" Sent")
+                else:
+                    print("[DEBUG] Packet Sending Failed, XBEE Closed")
+            except Exception as e:
+                print(f"[ERROR] Can't send packet: {e}")
