+++ main.py
@@ -1,5 +1,5 @@
 import app
-import newGCSXbee
+import GCSXbee
 import time
 
 if __name__ == "__main__":
@@ -18,7 +18,7 @@ if __name__ == "__main__":
     gui_app = app.App(TEAM_ID)
 
     # Create a telemetry handler object
-    telemetry_handler = newGCSXbee.TelemetryHandler(TEAM_ID, xbee_port=DEVICE_FILE, xbee_baudrate=BAUDRATE, xbee_target_mac_addr=XBEE_MAC_ADDR, press_csv_path=PRESSURE_FILE, log_csv_path=LOG_FILE)
+    telemetry_handler = GCSXbee.TelemetryHandler(TEAM_ID, xbee_port=DEVICE_FILE, xbee_baudrate=BAUDRATE, xbee_target_mac_addr=XBEE_MAC_ADDR, press_csv_path=PRESSURE_FILE, log_csv_path=LOG_FILE)
     
     telemetry_handler.start_telemetry()
     gui_app.mainloop()
